# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WayFarer is a public bus transportation booking server",
        "version": "1.0.0",
        "title": "Wayfarer API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "paths": {
...
      "/auth/signin": {
        "post": {
          "tags": [
            "auth"
          ],
          "summary": "Login user",
          "description": "Login of old users .",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "email, password",
              "description": "Logged in user object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Signin"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation"
            },
            "400": {
              "description": "User does not exist"
            }
          }
        }
      },
...
    },
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "definitions": {
...
      "Signin": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Auth"
        }
      },
...
    },
    "schemes": [
      "https",
      "http"
    ],
    "host":"https://wayfarer-bus-booking.herokuapp.com",
    "basePath": "/api"
  }